style: se mejora el backgroud de imagenes de home
